Move History serializers to ScodecSerialize
