+++ rspace/src/main/scala/coop/rchain/rspace/history/History.scala
@@ -4,22 +4,42 @@ import coop.rchain.crypto.codec.Base16
 import coop.rchain.rspace.Blake2b256Hash
 import coop.rchain.rspace.history.History._
 import coop.rchain.rspace.serializers.ScodecSerialize.{RichAttempt, _}
-import coop.rchain.shared.Serialize._
 import scodec.bits.{BitVector, ByteVector}
-import scodec.codecs.{discriminated, provide, uint, uint2, vectorOfN}
-import scodec.{Attempt, Codec}
 
+/**
+  * History definition represents key-value API for RSpace tuple space
+  *
+  * [[History]] contains only references to data stored on keys ([[KeyPath]]).
+  *
+  * [[ColdStoreInstances.ColdKeyValueStore]] holds full data referenced by [[LeafPointer]] in [[History]].
+  */
 trait History[F[_]] {
+
+  /**
+    * Insert/update/delete operations on the underlying Merkle tree (key-value store)
+    */
   def process(actions: List[HistoryAction]): F[History[F]]
-  def root: Blake2b256Hash
+
+  /**
+    * Read operation on the Merkle tree
+    */
   def find(key: KeyPath): F[(TriePointer, Vector[Trie])]
+
+  /**
+    * Get the root of the Merkle tree
+    */
+  def root: Blake2b256Hash
+
+  /**
+    * Returns History with specified with root pointer
+    */
   def reset(root: Blake2b256Hash): History[F]
 }
 
 object History {
 
   val emptyRoot: Trie               = EmptyTrie
-  private[this] def encodeEmptyRoot = Trie.codecTrie.encode(emptyRoot).get.toByteVector
+  private[this] def encodeEmptyRoot = codecTrie.encode(emptyRoot).get.toByteVector
   val emptyRootHash: Blake2b256Hash = Blake2b256Hash.create(encodeEmptyRoot)
 
   // this mapping is kept explicit on purpose
@@ -35,17 +55,21 @@ object History {
   def commonPrefix(l: KeyPath, r: KeyPath): KeyPath =
     (l.view, r.view).zipped.takeWhile { case (ll, rr) => ll == rr }.map(_._1).toSeq
 
-  def skip(pb: PointerBlock, affix: KeyPath): Trie =
-    Skip(ByteVector(affix), NodePointer(pb.hash))
-
   type KeyPath = Seq[Byte]
 }
 
+/*
+ * Type definitions for Merkle Trie implementation (History)
+ */
+
 sealed trait Trie
+
 sealed trait NonEmptyTrie extends Trie
-case object EmptyTrie     extends Trie
+
+case object EmptyTrie extends Trie
+
 final case class Skip(affix: ByteVector, ptr: ValuePointer) extends NonEmptyTrie {
-  lazy val encoded: BitVector = Trie.codecSkip.encode(this).get
+  lazy val encoded: BitVector = codecSkip.encode(this).get
 
   lazy val hash: Blake2b256Hash = Blake2b256Hash.create(encoded.toByteVector)
 
@@ -61,7 +85,7 @@ final case class PointerBlock private (toVector: Vector[TriePointer]) extends No
 
   def countNonEmpty: Int = toVector.count(_ != EmptyPointer)
 
-  lazy val encoded: BitVector = PointerBlock.codecPointerBlock.encode(this).get
+  lazy val encoded: BitVector = codecPointerBlock.encode(this).get
 
   lazy val hash: Blake2b256Hash = Blake2b256Hash.create(encoded.toByteVector)
 
@@ -76,97 +100,55 @@ final case class PointerBlock private (toVector: Vector[TriePointer]) extends No
   }
 }
 
-sealed trait TriePointer
-sealed trait NonEmptyTriePointer extends TriePointer {
-  def hash: Blake2b256Hash
-}
-case object EmptyPointer  extends TriePointer
-sealed trait ValuePointer extends NonEmptyTriePointer
-
-final case class LeafPointer(hash: Blake2b256Hash) extends ValuePointer
-final case class SkipPointer(hash: Blake2b256Hash) extends NonEmptyTriePointer
-final case class NodePointer(hash: Blake2b256Hash) extends ValuePointer
-
 object Trie {
+
+  /**
+    * Creates hash of Merkle Trie
+    *
+    * TODO: Fix encoding to use codec for the whole [[Trie]] and not for specific inherited variant.
+    */
   def hash(trie: Trie): Blake2b256Hash =
     trie match {
       case pb: PointerBlock  => pb.hash
       case s: Skip           => s.hash
       case _: EmptyTrie.type => History.emptyRootHash
     }
-
-  val codecSkip: Codec[Skip] = (codecByteVector :: codecTrieValuePointer).as[Skip]
-
-  val memoizingSkipCodec: Codec[Skip] =
-    Codec.apply((s: Skip) => Attempt.successful(s.encoded), codecSkip.decode)
-
-  val memoizingPointerBlockCodec: Codec[PointerBlock] =
-    Codec.apply(
-      (s: PointerBlock) => Attempt.successful(s.encoded),
-      PointerBlock.codecPointerBlock.decode
-    )
-
-  val codecTrie: Codec[Trie] =
-    discriminated[Trie]
-      .by(uint2)
-      .subcaseP(0) {
-        case e: EmptyTrie.type => e
-      }(provide(EmptyTrie))
-      .subcaseP(1) {
-        case s: Skip => s
-      }(memoizingSkipCodec)
-      .subcaseP(2) {
-        case pb: PointerBlock => pb
-      }(memoizingPointerBlockCodec)
-
-  implicit def codecTriePointer: Codec[TriePointer] =
-    discriminated[TriePointer]
-      .by(uint2)
-      .subcaseP(0) {
-        case p: EmptyPointer.type => p
-      }(provide(EmptyPointer))
-      .subcaseP(1) {
-        case p: LeafPointer => p
-      }(Blake2b256Hash.codecWithBytesStringBlake2b256Hash.as[LeafPointer])
-      .subcaseP(2) {
-        case p: SkipPointer => p
-      }(Blake2b256Hash.codecWithBytesStringBlake2b256Hash.as[SkipPointer])
-      .subcaseP(3) {
-        case p: NodePointer => p
-      }(Blake2b256Hash.codecWithBytesStringBlake2b256Hash.as[NodePointer])
-
-  implicit def codecTrieValuePointer: Codec[ValuePointer] =
-    discriminated[ValuePointer]
-      .by(uint(1))
-      .subcaseP(0) {
-        case p: LeafPointer => p
-      }(Blake2b256Hash.codecWithBytesStringBlake2b256Hash.as[LeafPointer])
-      .subcaseP(1) {
-        case p: NodePointer => p
-      }(Blake2b256Hash.codecWithBytesStringBlake2b256Hash.as[NodePointer])
-
 }
 
 object PointerBlock {
   val length = 256
 
-  def create(): PointerBlock = new PointerBlock(Vector.fill(length)(EmptyPointer))
+  val empty: PointerBlock = new PointerBlock(Vector.fill(length)(EmptyPointer))
 
-  def create(first: (Int, TriePointer)): PointerBlock =
-    PointerBlock.create().updated(List(first))
-
-  def create(first: (Int, TriePointer), second: (Int, TriePointer)): PointerBlock =
-    PointerBlock.create().updated(List(first, second))
-
-  // consider using zlib
-  implicit val codecPointerBlock: Codec[PointerBlock] =
-    vectorOfN(
-      provide(length),
-      Trie.codecTriePointer
-    ).as[PointerBlock]
+  def apply(first: (Int, TriePointer), second: (Int, TriePointer)): PointerBlock =
+    PointerBlock.empty.updated(List(first, second))
 
   def unapply(arg: PointerBlock): Option[Vector[TriePointer]] = Option(arg.toVector)
 }
+
+/*
+ * Trie pointer definitions
+ */
+
+sealed trait TriePointer
+
+sealed trait NonEmptyTriePointer extends TriePointer {
+  def hash: Blake2b256Hash
+}
+
+case object EmptyPointer extends TriePointer
+
+sealed trait ValuePointer extends NonEmptyTriePointer
+
+final case class LeafPointer(hash: Blake2b256Hash) extends ValuePointer
+
+final case class SkipPointer(hash: Blake2b256Hash) extends NonEmptyTriePointer
+
+final case class NodePointer(hash: Blake2b256Hash) extends ValuePointer
+
+/*
+ * Trie path used as a helper for traversal in History implementation
+ */
 
 final case class TriePath(nodes: Vector[Trie], conflicting: Option[Trie], edges: KeyPath) {
   def append(affix: KeyPath, t: Trie): TriePath =
