+++ rspace/src/main/scala/coop/rchain/rspace/state/RSpaceImporter.scala
@@ -4,7 +4,6 @@ import cats.effect._
 import cats.syntax.all._
 import coop.rchain.rspace.Blake2b256Hash
 import coop.rchain.rspace.history._
-import coop.rchain.rspace.serializers.ScodecSerialize
 import coop.rchain.rspace.serializers.ScodecSerialize._
 import coop.rchain.shared.AttemptOps.{RichAttempt, _}
 import coop.rchain.shared.Serialize
@@ -45,7 +44,7 @@ object RSpaceImporter {
     def raiseErrorEx[T](msg: String, cause: Throwable): F[T] =
       new StateValidationError(msg, cause).raiseError[F, T]
 
-    def decodeTrie(bytes: ByteVector): Trie = Trie.codecTrie.decodeValue(bytes.bits).get
+    def decodeTrie(bytes: ByteVector): Trie = codecTrie.decodeValue(bytes.bits).get
 
     def decodeData(bytes: ByteVector) =
       ColdStoreInstances.codecPersistedData.decodeValue(bytes.bits).toEitherThrowable
@@ -103,7 +102,7 @@ object RSpaceImporter {
                       .delay {
                         persistedData match {
                           case JoinsLeaf(bytes)         => decodeJoins[C](bytes)
-                          case DataLeaf(bytes)          => ScodecSerialize.decodeData[A](bytes)
+                          case DataLeaf(bytes)          => decodeDatums[A](bytes)
                           case ContinuationsLeaf(bytes) => decodeContinuations[P, K](bytes)
                         }
                       }
